+++ packages/tailwind/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.4.0](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/tailwind@0.3.4...@kaizen/tailwind@0.4.0) (2023-02-14)
+
+
+### Features
+
+* **tailwind:** Add spacing token for 6px/.375 rem ([#3269](https://github.com/cultureamp/kaizen-design-system/issues/3269)) ([7020545](https://github.com/cultureamp/kaizen-design-system/commit/7020545a294bb072ca151e08ef8d1c628b32b4dd))
+
+
+
+
+
+## [0.3.4](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/tailwind@0.3.3...@kaizen/tailwind@0.3.4) (2023-02-13)
+
+**Note:** Version bump only for package @kaizen/tailwind
+
+
+
+
+
+## [0.3.3](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/tailwind@0.3.2...@kaizen/tailwind@0.3.3) (2023-02-12)
+
+**Note:** Version bump only for package @kaizen/tailwind
+
+
+
+
+
 ## [0.3.2](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/tailwind@0.3.1...@kaizen/tailwind@0.3.2) (2023-01-31)
 
 **Note:** Version bump only for package @kaizen/tailwind